Remove all buildkite references (#1444)
What changed?
Remove all deprecated references to buildkite.

Why?
Because we are migrating to only use Github Actions.

How did you test it?
I am testing using the Github Actions pipeline.

Potential risks
Removing it could make it so that a push/PR passes when it shouldn't (because the Github Actions failed to catch it).
---------

Co-authored-by: Mandy Chen <mandy.chen@uber.com>
Co-authored-by: Shijie Sheng <shengs@uber.com>