Possible fix for a goroutine deadlock issue
Thanks to some internal tooling to look for deadlocked goroutines, this test:
> appears to go.uber.org/cadence/internal.(*WorkflowTestSuiteUnitTest).Test_ContextMisuse
appears to trigger a goroutine leak.

Since it's not doing anything *impossible* for users to do (it's just using the wrong context arg),
it seems like this implies we've got incorrect goroutine shutdown code.

I'm not yet fully confident that this is The Fix and that it does not cause other issues, but it seems pretty likely so far.
I don't know if this would explain the in-production leaks we've seen in the past but not reproduced (those seem to be query-triggered somehow), but it *does* seem quite likely that this has leaked somewhere.  Just hopefully also caused errors that pointed to bad code which led to a fix.